Add files via upload
I've added the French language file. In order to use it, create the "lang" folder next to your launcher if you don't already have it and place this file inside, then start your launcher, go to Settings and select the French language. After a restart, your UI will be in French.