Added the Chinese language file
I've added the Mandarin Chinese language file. This file uses the simplified Han characters of Mandarin Chinese after 1952, which is most widely used today. In order to use it, create the "lang" folder next to your launcher if you don't already have it and place this file inside, then start your launcher, go to Settings and select the Chinese language. After a restart, your UI will be in Chinese. Why the hell does that sound like an insult? I didn't mean it that way.